Promote stride parameters to int64_t for swizzle and format convert C API
Change src_row_bytes/dst_row_bytes from uint32_t to int64_t in:
- nozzle_swizzle_channels
- nozzle_widen_uint16_to_uint32
- nozzle_convert_uint32_to_float32
- nozzle_widen_half_to_float

This matches nozzle_fill_opaque_alpha_channel which already uses int64_t
and NozzleMappedPixels.row_stride_bytes which is int64_t.

Validation: negative strides rejected (these functions don't support them).
SIMD/vImage dispatch paths keep uint32_t internally (validated downcast).
C++ public API signatures updated to match.

Tests: negative src/dst row_bytes rejection tests added.

+++ include/nozzle/channel_swizzle.hpp
@@ -26,7 +26,7 @@ constexpr channel_permute permute_rgba_to_argb{{3, 0, 1, 2}};
 Result<void> swizzle_channels(
 	const void *src, void *dst,
 	uint32_t width, uint32_t height,
-	uint32_t src_row_bytes, uint32_t dst_row_bytes,
+	int64_t src_row_bytes, int64_t dst_row_bytes,
 	texture_format format,
 	channel_permute permute);
