+++ include/nozzle/format_convert.hpp
@@ -13,7 +13,7 @@ namespace nozzle {
 Result<void> widen_uint16_to_uint32(
 	const void *src, void *dst,
 	uint32_t width, uint32_t height,
-	uint32_t src_row_bytes, uint32_t dst_row_bytes,
+	int64_t src_row_bytes, int64_t dst_row_bytes,
 	uint32_t channels);
 
 // Convert uint32 buffer to float32 buffer.
@@ -22,7 +22,7 @@ Result<void> widen_uint16_to_uint32(
 Result<void> convert_uint32_to_float32(
 	const void *src, void *dst,
 	uint32_t width, uint32_t height,
-	uint32_t src_row_bytes, uint32_t dst_row_bytes,
+	int64_t src_row_bytes, int64_t dst_row_bytes,
 	uint32_t channels);
 
 // Widen half-float (IEEE 754-2008 binary16) buffer to float32 buffer.
@@ -31,7 +31,7 @@ Result<void> convert_uint32_to_float32(
 Result<void> widen_half_to_float(
 	const void *src, void *dst,
 	uint32_t width, uint32_t height,
-	uint32_t src_row_bytes, uint32_t dst_row_bytes,
+	int64_t src_row_bytes, int64_t dst_row_bytes,
 	uint32_t channels);
 
 // Fill the alpha channel of a 4-channel pixel buffer with the canonical opaque value.
