+++ src/c_api/nozzle_c.cpp
@@ -811,8 +811,8 @@ NozzleErrorCode nozzle_swizzle_channels(
     void *dst,
     uint32_t width,
     uint32_t height,
-    uint32_t src_row_bytes,
-    uint32_t dst_row_bytes,
+    int64_t src_row_bytes,
+    int64_t dst_row_bytes,
     NozzleTextureFormat format,
     const uint8_t permute_map[4]
 ) {
@@ -843,8 +843,8 @@ NozzleErrorCode nozzle_widen_uint16_to_uint32(
     void *dst,
     uint32_t width,
     uint32_t height,
-    uint32_t src_row_bytes,
-    uint32_t dst_row_bytes,
+    int64_t src_row_bytes,
+    int64_t dst_row_bytes,
     uint32_t channels
 ) {
     auto result = nozzle::widen_uint16_to_uint32(
@@ -857,8 +857,8 @@ NozzleErrorCode nozzle_convert_uint32_to_float32(
     void *dst,
     uint32_t width,
     uint32_t height,
-    uint32_t src_row_bytes,
-    uint32_t dst_row_bytes,
+    int64_t src_row_bytes,
+    int64_t dst_row_bytes,
     uint32_t channels
 ) {
     auto result = nozzle::convert_uint32_to_float32(
@@ -871,8 +871,8 @@ NozzleErrorCode nozzle_widen_half_to_float(
     void *dst,
     uint32_t width,
     uint32_t height,
-    uint32_t src_row_bytes,
-    uint32_t dst_row_bytes,
+    int64_t src_row_bytes,
+    int64_t dst_row_bytes,
     uint32_t channels
 ) {
     auto result = nozzle::widen_half_to_float(
